refactor(workflowexecution): Phase 6 Wave 2c - decompose reconcilePending, split oversized file
Adds characterization tests (pkg/workflowexecution/reconcile_pending_orchestration_test.go)
covering reconcilePending's orchestration branches that lacked direct unit
coverage despite the existing black-box Reconcile() suites: cooldown
enforcement (FedRAMP SC-5 DoS protection) and audit idempotency (SOC2 CC8.1
completeness). Written and confirmed green before refactoring, per AGENTS.md's
coverage-before-refactor mandate.

Decomposes reconcilePending (49 -> 7 cyclomatic) into 7 independently testable
steps: refetchFreshPendingWFE, validateAndAnnouncePendingSpec,
resolvePendingSchemaAndEngine, checkPendingCooldownOrBlock,
recordPendingSelectionAudit, createPendingExecutionResource, and
finalizePendingToRunning. Fixes a latent bug where checkPendingCooldownOrBlock
silently dropped a status-update error instead of propagating it.

Splits workflowexecution_controller.go (1,979 lines post-decomposition) into
workflowexecution_controller.go (core dispatcher + NewReconciler + ValidateSpec),
workflowexecution_pending.go, workflowexecution_status_marking.go,
workflowexecution_lifecycle.go, workflowexecution_collision.go, and
workflowexecution_catalog.go. All files now under the 700-line convention
threshold.

Per GO-ANTIPATTERN-AUDIT-2026-07-01 Wave 2 (issue #1520, see §7j-1).
