+++ internal/controller/workflowexecution/workflowexecution_controller.go
@@ -50,15 +50,10 @@ import (
 	tektonv1 "github.com/tektoncd/pipeline/pkg/apis/pipeline/v1"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
-	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/record"
-	"knative.dev/pkg/apis"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -71,15 +66,14 @@ import (
 
 	workflowexecutionv1alpha1 "github.com/jordigilh/kubernaut/api/workflowexecution/v1alpha1"
 	"github.com/jordigilh/kubernaut/pkg/audit"
+	dsvalidation "github.com/jordigilh/kubernaut/pkg/datastorage/validation"
 	"github.com/jordigilh/kubernaut/pkg/shared/events"
-	weconditions "github.com/jordigilh/kubernaut/pkg/workflowexecution"
+	"github.com/jordigilh/kubernaut/pkg/shared/k8serrors"
 	weaudit "github.com/jordigilh/kubernaut/pkg/workflowexecution/audit"
 	weclient "github.com/jordigilh/kubernaut/pkg/workflowexecution/client"
 	weexecutor "github.com/jordigilh/kubernaut/pkg/workflowexecution/executor"
 	"github.com/jordigilh/kubernaut/pkg/workflowexecution/metrics"
 	wephase "github.com/jordigilh/kubernaut/pkg/workflowexecution/phase"
-	dsvalidation "github.com/jordigilh/kubernaut/pkg/datastorage/validation"
-	"github.com/jordigilh/kubernaut/pkg/shared/k8serrors"
 	"github.com/jordigilh/kubernaut/pkg/workflowexecution/status"
 )
 
@@ -98,7 +92,6 @@ const (
 
 	// DefaultCooldownPeriod is the default time between workflow executions on same target
 	DefaultCooldownPeriod = 5 * time.Minute
-
 )
 
 // WorkflowExecutionReconciler reconciles a WorkflowExecution object
@@ -225,45 +218,6 @@ func NewReconciler(mgr ctrl.Manager, opts ReconcilerOptions) *WorkflowExecutionR
 	}
 }
 
-// resolveSchemaMetadata fetches all workflow catalog artifacts from DS in a
-// single call (F6: consolidation of 3 DS round-trips into 1).
-// Returns nil metadata (not an error) when no querier is configured or when
-// the DS call fails — graceful degradation preserves execution without
-// filtering, engine config, or dependency validation.
-// Returns a hard error only when dependency validation explicitly fails.
-func (r *WorkflowExecutionReconciler) resolveSchemaMetadata(ctx context.Context, wfe *workflowexecutionv1alpha1.WorkflowExecution) (*weclient.SchemaMetadata, weexecutor.CreateOptions, error) {
-	opts := weexecutor.CreateOptions{}
-	if r.WorkflowQuerier == nil {
-		return nil, opts, nil
-	}
-
-	logger := log.FromContext(ctx)
-	meta, err := r.WorkflowQuerier.GetWorkflowSchemaMetadata(ctx, wfe.Spec.WorkflowRef.WorkflowID)
-	if err != nil {
-		logger.Error(err, "Failed to fetch workflow schema metadata from DS (non-fatal, continuing without filtering)",
-			"workflowID", wfe.Spec.WorkflowRef.WorkflowID)
-		if r.Metrics != nil {
-			r.Metrics.SchemaFetchFailures.Inc()
-		}
-		return nil, opts, nil
-	}
-	if meta == nil {
-		return nil, opts, nil
-	}
-
-	if meta.Dependencies != nil && r.DependencyValidator != nil {
-		if valErr := r.DependencyValidator.ValidateDependencies(ctx, r.ExecutionNamespace, meta.Dependencies); valErr != nil {
-			logger.Error(valErr, "Workflow dependency validation failed",
-				"workflowID", wfe.Spec.WorkflowRef.WorkflowID)
-			return meta, opts, valErr
-		}
-	}
-
-	opts.Dependencies = meta.Dependencies
-	opts.DeclaredParameterNames = meta.DeclaredParameterNames
-	return meta, opts, nil
-}
-
 // ========================================
 // RBAC Markers
 // ========================================
@@ -348,655 +302,6 @@ func (r *WorkflowExecutionReconciler) Reconcile(ctx context.Context, req ctrl.Re
 	}
 
 	// Should never reach here
-	return ctrl.Result{}, nil
-}
-
-// ========================================
-// reconcilePending - Handle Pending phase
-// V1.0: Pure execution logic, RO handles all routing (DD-RO-002)
-// ========================================
-func (r *WorkflowExecutionReconciler) reconcilePending(ctx context.Context, wfe *workflowexecutionv1alpha1.WorkflowExecution) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-	logger.Info("Reconciling Pending phase")
-
-	// ========================================
-	// DD-STATUS-001: Re-read WFE from API server to bypass informer cache.
-	// Prevents race conditions where concurrent reconciles read stale data:
-	// - F1: ExecutionRef not yet cached → misses external deletion detection
-	// - F2: Phase not yet cached → re-enters reconcilePending → duplicate audit events
-	// ========================================
-	freshWFE := &workflowexecutionv1alpha1.WorkflowExecution{}
-	if err := r.APIReader.Get(ctx, client.ObjectKeyFromObject(wfe), freshWFE); err != nil {
-		if apierrors.IsNotFound(err) {
-			return ctrl.Result{}, nil // WFE was deleted
-		}
-		return ctrl.Result{}, fmt.Errorf("failed to re-read WFE from API server: %w", err)
-	}
-	// If the fresh WFE has already progressed past Pending, requeue to let the
-	// main Reconcile re-route based on the updated phase.
-	if freshWFE.Status.Phase != "" && freshWFE.Status.Phase != workflowexecutionv1alpha1.PhasePending {
-		logger.Info("WFE already progressed past Pending (informer cache was stale), requeueing",
-			"freshPhase", freshWFE.Status.Phase)
-		return ctrl.Result{Requeue: true}, nil
-	}
-	// Use fresh data for the remainder of this reconcile
-	*wfe = *freshWFE
-
-	// V1.0: No routing logic - RO makes ALL routing decisions before creating WFE
-	// If WFE exists, execute it. RO already checked routing.
-
-	// ========================================
-	// Step 1: Validate spec (prevent malformed PipelineRuns)
-	// ========================================
-	if err := r.ValidateSpec(wfe); err != nil {
-		logger.Error(err, "Spec validation failed")
-		// DD-EVENT-001 v1.1: Emit WorkflowValidationFailed event (P2: Decision Point)
-		r.Recorder.Event(wfe, corev1.EventTypeWarning, events.EventReasonWorkflowValidationFailed,
-			fmt.Sprintf("Spec validation failed: %v", err))
-		// Mark as Failed with ConfigurationError reason
-		// This is a pre-execution failure (wasExecutionFailure: false)
-		if markErr := r.MarkFailedWithReason(ctx, wfe, "ConfigurationError", err.Error()); markErr != nil {
-			return ctrl.Result{}, markErr
-		}
-		return ctrl.Result{}, nil
-	}
-
-	// DD-EVENT-001 v1.1: Emit WorkflowValidated event (P2: Decision Point)
-	r.Recorder.Event(wfe, corev1.EventTypeNormal, events.EventReasonWorkflowValidated,
-		fmt.Sprintf("Workflow spec validated: %s (target: %s)", wfe.Spec.WorkflowRef.WorkflowID, wfe.Spec.TargetResource))
-
-	// ========================================
-	// Step 1.2: Resolve all workflow catalog metadata from DS in a single call (F6)
-	// Engine (#518), engineConfig (DD-WORKFLOW-017), dependencies (DD-WE-006),
-	// and declared parameter names (#243) are fetched together.
-	// ========================================
-	schemaMeta, createOpts, schemaErr := r.resolveSchemaMetadata(ctx, wfe)
-	if schemaErr != nil {
-		r.Recorder.Event(wfe, corev1.EventTypeWarning, events.EventReasonWorkflowValidationFailed,
-			fmt.Sprintf("Workflow dependency validation failed: %v", schemaErr))
-		markErr := r.MarkFailedWithReason(ctx, wfe, "ConfigurationError", schemaErr.Error())
-		return ctrl.Result{}, markErr
-	}
-
-	// DD-WORKFLOW-017: Resolve engineConfig from schema before catalog (schema takes precedence).
-	if wfe.Spec.WorkflowRef.EngineConfig == nil && schemaMeta != nil && schemaMeta.EngineConfig != nil {
-		logger.Info("Resolved engineConfig from DS schema metadata",
-			"workflowID", wfe.Spec.WorkflowRef.WorkflowID)
-		wfe.Spec.WorkflowRef.EngineConfig = &apiextensionsv1.JSON{Raw: schemaMeta.EngineConfig}
-	}
-
-	// Step 1.3: Resolve execution bundle, SA, and engine from DS catalog (Issue #650).
-	// Called BEFORE setting engine from schema so the idempotency guard in
-	// resolveWorkflowCatalog doesn't skip the SA and bundle resolution that
-	// only the catalog provides.
-	if _, catalogErr := r.resolveWorkflowCatalog(ctx, wfe); catalogErr != nil {
-		logger.Error(catalogErr, "Failed to resolve workflow catalog from DS (non-fatal)")
-	}
-
-	// Resolve execution engine: prefer catalog (includes SA + bundle), schema fallback (Issue #518).
-	if wfe.Status.ExecutionEngine == "" && schemaMeta != nil && schemaMeta.Engine != "" {
-		wfe.Status.ExecutionEngine = schemaMeta.Engine
-	}
-	engine := wfe.Status.ExecutionEngine
-	if engine == "" {
-		engineErr := fmt.Errorf("no engine defined in remediation workflow %s", wfe.Spec.WorkflowRef.WorkflowID)
-		if schemaMeta != nil && schemaMeta.WorkflowName != "" {
-			engineErr = fmt.Errorf("no engine defined in remediation workflow %s - %s",
-				schemaMeta.WorkflowName, wfe.Spec.WorkflowRef.WorkflowID)
-		}
-		logger.Error(engineErr, "Execution engine resolution failed")
-		r.Recorder.Event(wfe, corev1.EventTypeWarning, events.EventReasonWorkflowValidationFailed,
-			fmt.Sprintf("Workflow catalog resolution failed: %v", engineErr))
-		if markErr := r.MarkFailedWithReason(ctx, wfe, "ConfigurationError", engineErr.Error()); markErr != nil {
-			return ctrl.Result{}, markErr
-		}
-		return ctrl.Result{}, nil
-	}
-	logger.Info("Resolved execution engine from catalog", "engine", engine, "workflowID", wfe.Spec.WorkflowRef.WorkflowID)
-
-	// ========================================
-	// Step 1.5: Check if cooldown is active for target resource (BR-WE-009)
-	// BUGFIX: Was only tracked in terminal phase, not enforced during pending
-	// ========================================
-	currentWFEKey := fmt.Sprintf("%s/%s", wfe.Namespace, wfe.Name)
-	if remaining, active := r.CheckCooldownActive(ctx, wfe.Spec.TargetResource, currentWFEKey); active {
-		logger.Info("Blocking execution due to active cooldown",
-			"targetResource", wfe.Spec.TargetResource,
-			"remaining", remaining,
-		)
-		// Ensure phase is set to Pending if not already set (P0: Phase State Machine)
-		if wfe.Status.Phase == "" || wfe.Status.Phase != workflowexecutionv1alpha1.PhasePending {
-			if err := r.PhaseManager.TransitionTo(wfe, wephase.Pending); err != nil {
-				return ctrl.Result{}, fmt.Errorf("failed to transition to Pending during cooldown: %w", err)
-			}
-			if err := r.Status().Update(ctx, wfe); err != nil {
-				return ctrl.Result{}, fmt.Errorf("failed to update phase to Pending during cooldown: %w", err)
-			}
-		}
-		// DD-EVENT-001 v1.1: Emit CooldownActive event (P2: Decision Point)
-		r.Recorder.Event(wfe, corev1.EventTypeNormal, events.EventReasonCooldownActive,
-			fmt.Sprintf("Execution deferred: cooldown active for %s (remaining: %s)", wfe.Spec.TargetResource, remaining.Round(time.Second)))
-
-		// Stay in Pending, requeue after cooldown expires
-		return ctrl.Result{RequeueAfter: remaining}, nil
-	}
-
-	// ========================================
-	// Gap #5: Record workflow selection audit event (BR-AUDIT-005)
-	// Emitted AFTER validation, BEFORE execution resource creation
-	// IDEMPOTENCY: Only emit once - skip if execution resource already exists
-	// ========================================
-	// DD-STATUS-001: Use APIReader to bypass informer cache for existence check.
-	// Prevents duplicate audit events when concurrent reconciles don't yet see a recently-created resource in cache.
-	resourceName := weexecutor.ExecutionResourceName(wfe.Spec.TargetResource)
-	resourceExists := false
-	switch wfe.Status.ExecutionEngine {
-	case "job":
-		existingJob := &batchv1.Job{}
-		err := r.APIReader.Get(ctx, client.ObjectKey{Name: resourceName, Namespace: r.ExecutionNamespace}, existingJob)
-		if err == nil {
-			resourceExists = true
-		} else if apierrors.IsNotFound(err) && wfe.Status.ExecutionRef != nil {
-			logger.Error(err, "Execution resource not found - deleted externally during Pending phase",
-				"engine", wfe.Status.ExecutionEngine)
-			return r.MarkFailed(ctx, wfe, nil)
-		}
-	case "tekton":
-		existingPR := &tektonv1.PipelineRun{}
-		err := r.APIReader.Get(ctx, client.ObjectKey{Name: resourceName, Namespace: r.ExecutionNamespace}, existingPR)
-		if err == nil {
-			resourceExists = true
-		} else if apierrors.IsNotFound(err) && wfe.Status.ExecutionRef != nil {
-			logger.Error(err, "Execution resource not found - deleted externally during Pending phase",
-				"engine", wfe.Status.ExecutionEngine)
-			return r.MarkFailed(ctx, wfe, nil)
-		}
-	case "ansible":
-		// AWX manages jobs externally — no K8s resource to check.
-		// Use ExecutionRef as the idempotency signal.
-		resourceExists = wfe.Status.ExecutionRef != nil
-	default:
-		logger.Info("Unknown engine for existence check, skipping audit guard",
-			"engine", wfe.Status.ExecutionEngine)
-	}
-
-	if !resourceExists {
-		workflowName := ""
-		if schemaMeta != nil {
-			workflowName = schemaMeta.WorkflowName
-		}
-		if err := r.AuditManager.RecordWorkflowSelectionCompleted(ctx, wfe, workflowName); err != nil {
-			logger.V(1).Info("Failed to record workflow.selection.completed audit event", "error", err)
-		}
-	} else {
-		logger.V(2).Info("Skipping workflow.selection.completed audit event - execution resource already exists",
-			"resource", resourceName, "engine", wfe.Status.ExecutionEngine)
-	}
-
-	// ========================================
-	// Step 2: Create execution resource via executor dispatch (BR-WE-014)
-	// ========================================
-	exec, err := r.ExecutorRegistry.Get(wfe.Status.ExecutionEngine)
-	if err != nil {
-		// Issue #868: Provide actionable guidance for unavailable engines
-		guidance := engineGuidance(wfe.Status.ExecutionEngine)
-		msg := fmt.Sprintf("execution engine %q is not available -- %s", wfe.Status.ExecutionEngine, guidance)
-		logger.Error(err, "Unsupported execution engine", "engine", wfe.Status.ExecutionEngine)
-		r.Recorder.Event(wfe, corev1.EventTypeWarning, events.EventReasonWorkflowValidationFailed, msg)
-		markErr := r.MarkFailedWithReason(ctx, wfe, "UnsupportedEngine", msg)
-		return ctrl.Result{}, markErr
-	}
-
-	logger.Info("Creating execution resource",
-		"engine", wfe.Status.ExecutionEngine,
-		"resource", resourceName,
-		"namespace", r.ExecutionNamespace,
-	)
-
-	// DD-WE-006 / F6: Dependencies, parameter names, and dependency validation
-	// already resolved by resolveSchemaMetadata above (createOpts is populated).
-
-	createResult, createErr := exec.Create(ctx, wfe, r.ExecutionNamespace, createOpts)
-	if createErr != nil {
-		if apierrors.IsAlreadyExists(createErr) {
-			// DD-WE-003 Layer 2: Execution-time collision handling
-			if wfe.Status.ExecutionEngine == "tekton" {
-				return r.HandleAlreadyExists(ctx, wfe, resourceName, createErr)
-			}
-			// Issue #374 / DD-WE-003: Pre-execution cleanup of completed Jobs.
-			// If the existing Job is in a terminal state (Succeeded/Failed), clean it up
-			// and retry creation. If still running, the lock is valid -- fail the WFE.
-			if wfe.Status.ExecutionEngine == "job" {
-				retryResult, handled, requeueForGC, originalWFE := r.handleJobAlreadyExists(ctx, exec, wfe, resourceName, createOpts)
-				if handled {
-					createResult = retryResult
-					createErr = nil
-				} else if requeueForGC {
-					logger.Info("Requeuing for Job GC completion (Issue #383)", "resource", resourceName)
-					return ctrl.Result{RequeueAfter: 500 * time.Millisecond}, nil
-				} else if originalWFE != "" {
-					// Issue #190: Valid lock owned by another WFE — classify as Deduplicated.
-					markErr := r.MarkFailedAsDeduplicated(ctx, wfe, originalWFE)
-					return ctrl.Result{}, markErr
-				}
-			}
-			if createErr != nil {
-				markErr := r.MarkFailedWithReason(ctx, wfe, "Unknown",
-					fmt.Sprintf("Execution resource %s already exists (target resource locked)", resourceName))
-				return ctrl.Result{}, markErr
-			}
-		} else {
-			logger.Error(createErr, "Failed to create execution resource", "engine", wfe.Status.ExecutionEngine)
-			markErr := r.MarkFailedWithReason(ctx, wfe, "Unknown",
-				fmt.Sprintf("Failed to create %s execution resource: %v", wfe.Status.ExecutionEngine, createErr))
-			return ctrl.Result{}, markErr
-		}
-	}
-
-	createdName := createResult.ResourceName
-
-	// Restore in-memory status fields that may have been lost if Create() performed
-	// an internal status update (e.g., storeCredentialIDs). These fields were set by
-	// resolveWorkflowCatalog but not yet persisted to the API server.
-	wfe.Status.ExecutionEngine = engine
-
-	// Issue #501: Process warnings from CreateResult (e.g., TokenTTL issues)
-	for _, w := range createResult.Warnings {
-		meta.SetStatusCondition(&wfe.Status.Conditions, metav1.Condition{
-			Type:               w.Type,
-			Status:             metav1.ConditionTrue,
-			Reason:             w.Reason,
-			Message:            w.Message,
-			ObservedGeneration: wfe.Generation,
-		})
-		r.Recorder.Event(wfe, corev1.EventTypeWarning, w.Reason, w.Message)
-		logger.Info("CreateResult warning applied", "type", w.Type, "reason", w.Reason)
-	}
-
-	// ========================================
-	// Gap #6: Record execution workflow started audit event (BR-AUDIT-005)
-	// ========================================
-	if err := r.AuditManager.RecordExecutionWorkflowStarted(ctx, wfe, createdName, r.ExecutionNamespace); err != nil {
-		logger.V(1).Info("Failed to record execution.workflow.started audit event", "error", err)
-		weconditions.SetAuditRecorded(wfe, false,
-			weconditions.ReasonAuditFailed,
-			fmt.Sprintf("Failed to record audit event: %v", err))
-	} else {
-		weconditions.SetAuditRecorded(wfe, true,
-			weconditions.ReasonAuditSucceeded,
-			"Audit event execution.workflow.started recorded to DataStorage")
-	}
-
-	// ========================================
-	// BR-WE-006: Set ExecutionCreated condition
-	// ========================================
-	weconditions.SetExecutionCreated(wfe, true,
-		weconditions.ReasonExecutionCreated,
-		fmt.Sprintf("%s execution resource %s created in %s namespace",
-			wfe.Status.ExecutionEngine, createdName, r.ExecutionNamespace))
-
-	// ========================================
-	// Step 3: Prepare status update to Running (P0: Phase State Machine)
-	// ========================================
-	now := metav1.Now()
-	if err := r.PhaseManager.TransitionTo(wfe, wephase.Running); err != nil {
-		return ctrl.Result{}, fmt.Errorf("failed to transition to Running: %w", err)
-	}
-	wfe.Status.StartTime = &now
-	wfe.Status.ExecutionRef = &corev1.LocalObjectReference{
-		Name: createdName,
-	}
-
-	// Single atomic status update with all changes
-	if err := r.updateStatus(ctx, wfe, "Running with conditions"); err != nil {
-		return ctrl.Result{}, err
-	}
-
-	r.Recorder.Event(wfe, corev1.EventTypeNormal, events.EventReasonExecutionCreated,
-		fmt.Sprintf("Created %s execution resource %s/%s", wfe.Status.ExecutionEngine, r.ExecutionNamespace, createdName))
-
-	// DD-EVENT-001 v1.1: PhaseTransition breadcrumb for Pending → Running
-	r.emitPhaseTransition(wfe, "Pending", "Running")
-
-	// Requeue to check execution status
-	return ctrl.Result{RequeueAfter: 10 * time.Second}, nil
-}
-
-// ========================================
-// reconcileRunning - Handle Running phase
-// Day 5: Status synchronization
-// ========================================
-func (r *WorkflowExecutionReconciler) reconcileRunning(ctx context.Context, wfe *workflowexecutionv1alpha1.WorkflowExecution) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-
-	// Issue #518: Lazy resolution for pre-migration WFEs that lack status.executionEngine.
-	if _, engineErr := r.resolveExecutionEngine(ctx, wfe); engineErr != nil {
-		logger.Error(engineErr, "Failed to resolve execution engine during Running phase")
-		return r.MarkFailed(ctx, wfe, nil)
-	}
-	logger.Info("Reconciling Running phase", "engine", wfe.Status.ExecutionEngine)
-
-	// ========================================
-	// Step 1: Get execution status via executor dispatch (BR-WE-014)
-	// ========================================
-	exec, err := r.ExecutorRegistry.Get(wfe.Status.ExecutionEngine)
-	if err != nil {
-		logger.Error(err, "Unsupported execution engine during Running phase")
-		return r.MarkFailed(ctx, wfe, nil)
-	}
-
-	result, getErr := exec.GetStatus(ctx, wfe, r.ExecutionNamespace)
-	if getErr != nil {
-		if apierrors.IsNotFound(getErr) {
-			logger.Error(getErr, "Execution resource not found - deleted externally",
-				"engine", wfe.Status.ExecutionEngine)
-			return r.MarkFailed(ctx, wfe, nil)
-		}
-		return ctrl.Result{}, getErr
-	}
-
-	// ========================================
-	// Step 2: Map execution result to WFE phase
-	// ExecutionStatus is passed into Mark* methods to survive AtomicStatusUpdate refetch
-	// ========================================
-	switch result.Phase {
-	case workflowexecutionv1alpha1.PhaseCompleted:
-		logger.Info("Execution succeeded", "engine", wfe.Status.ExecutionEngine)
-		if wfe.Status.ExecutionEngine == "tekton" {
-			var pr tektonv1.PipelineRun
-			if err := r.Get(ctx, client.ObjectKey{
-				Name:      wfe.Status.ExecutionRef.Name,
-				Namespace: r.ExecutionNamespace,
-			}, &pr); err != nil {
-				return r.MarkCompleted(ctx, wfe, nil, result.Summary)
-			}
-			return r.MarkCompleted(ctx, wfe, &pr, result.Summary)
-		}
-		return r.MarkCompleted(ctx, wfe, nil, result.Summary)
-
-	case workflowexecutionv1alpha1.PhaseFailed:
-		logger.Info("Execution failed", "engine", wfe.Status.ExecutionEngine, "reason", result.Reason)
-		if wfe.Status.ExecutionEngine == "tekton" {
-			var pr tektonv1.PipelineRun
-			if err := r.Get(ctx, client.ObjectKey{
-				Name:      wfe.Status.ExecutionRef.Name,
-				Namespace: r.ExecutionNamespace,
-			}, &pr); err != nil {
-				return r.MarkFailed(ctx, wfe, nil, result.Summary)
-			}
-			return r.MarkFailed(ctx, wfe, &pr, result.Summary)
-		}
-		return r.MarkFailed(ctx, wfe, nil, result.Summary)
-
-	default:
-		// Still running - update conditions and requeue
-		logger.V(1).Info("Execution still running", "reason", result.Reason, "engine", wfe.Status.ExecutionEngine)
-		weconditions.SetExecutionRunning(wfe, true,
-			weconditions.ReasonExecutionStarted,
-			fmt.Sprintf("Execution running (%s: %s)", wfe.Status.ExecutionEngine, result.Reason))
-	}
-
-	// ========================================
-	// Step 4: Update status with current progress
-	// ========================================
-	if err := r.updateStatus(ctx, wfe, "current progress"); err != nil {
-		return ctrl.Result{}, err
-	}
-
-	return ctrl.Result{RequeueAfter: 10 * time.Second}, nil
-}
-
-// ========================================
-// ReconcileTerminal - Handle Completed/Failed phases
-// Day 6: Cooldown enforcement and cleanup
-// DD-WE-003: Lock Persistence (Deterministic Name)
-// ========================================
-func (r *WorkflowExecutionReconciler) ReconcileTerminal(ctx context.Context, wfe *workflowexecutionv1alpha1.WorkflowExecution) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-
-	// Issue #518: Lazy resolution for pre-migration WFEs (non-fatal for terminal phase).
-	if _, engineErr := r.resolveExecutionEngine(ctx, wfe); engineErr != nil {
-		logger.V(1).Info("Could not resolve execution engine in terminal phase (non-fatal)", "error", engineErr)
-	}
-	logger.Info("Reconciling Terminal phase", "phase", wfe.Status.Phase)
-
-	// WE-H3: No completion time in terminal phase means we can't calculate cooldown,
-	// but we must still release execution resources. Set completionTime now so cleanup proceeds.
-	if wfe.Status.CompletionTime == nil {
-		logger.Info("No completion time set in terminal phase — setting now to unblock cleanup")
-		now := metav1.Now()
-		wfe.Status.CompletionTime = &now
-	}
-
-	// Get cooldown period (use default if not set)
-	cooldown := r.CooldownPeriod
-	if cooldown == 0 {
-		cooldown = DefaultCooldownPeriod
-	}
-
-	// Calculate elapsed time since completion
-	elapsed := time.Since(wfe.Status.CompletionTime.Time)
-
-	// Wait for cooldown before releasing lock
-	if elapsed < cooldown {
-		remaining := cooldown - elapsed
-		logger.V(1).Info("Waiting for cooldown",
-			"remaining", remaining,
-			"targetResource", wfe.Spec.TargetResource,
-		)
-		return ctrl.Result{RequeueAfter: remaining}, nil
-	}
-
-	// Cooldown expired - delete execution resource to release lock
-	// DD-WE-003: Use deterministic name for atomic locking
-	if r.ExecutorRegistry != nil {
-		exec, execErr := r.ExecutorRegistry.Get(wfe.Status.ExecutionEngine)
-		if execErr != nil {
-			logger.Error(execErr, "Unknown engine during cooldown cleanup, skipping",
-				"engine", wfe.Status.ExecutionEngine)
-		}
-		if exec != nil {
-			if err := exec.Cleanup(ctx, wfe, r.ExecutionNamespace); err != nil {
-				logger.Error(err, "Failed to cleanup execution resource after cooldown",
-					"engine", exec.Engine())
-				// DD-EVENT-001 v1.1: Emit CleanupFailed event (P4: Error Path)
-				r.Recorder.Event(wfe, corev1.EventTypeWarning, events.EventReasonCleanupFailed,
-					fmt.Sprintf("Cleanup failed after cooldown: %v", err))
-				return ctrl.Result{}, err
-			}
-			logger.Info("Lock released after cooldown",
-				"targetResource", wfe.Spec.TargetResource,
-				"engine", exec.Engine(),
-				"cooldownPeriod", cooldown,
-			)
-		}
-	} else {
-		// Fallback: inline Tekton cleanup when ExecutorRegistry is not configured
-		prName := weexecutor.ExecutionResourceName(wfe.Spec.TargetResource)
-		var existing tektonv1.PipelineRun
-		if err := r.Get(ctx, client.ObjectKey{Name: prName, Namespace: r.ExecutionNamespace}, &existing); err != nil {
-			if !apierrors.IsNotFound(err) {
-				logger.Error(err, "Failed to get PipelineRun for ownership check")
-				return ctrl.Result{}, err
-			}
-			// Already gone -- nothing to clean up
-		} else if existing.Labels["kubernaut.ai/workflow-execution"] == wfe.Name {
-			if err := r.Delete(ctx, &existing); err != nil && !apierrors.IsNotFound(err) {
-				logger.Error(err, "Failed to delete PipelineRun after cooldown")
-				return ctrl.Result{}, err
-			}
-		}
-		logger.Info("Lock released after cooldown",
-			"targetResource", wfe.Spec.TargetResource,
-			"cooldownPeriod", cooldown,
-		)
-	}
-
-	// Emit LockReleased event
-	r.Recorder.Event(wfe, corev1.EventTypeNormal, events.EventReasonLockReleased,
-		fmt.Sprintf("Lock released for %s after cooldown", wfe.Spec.TargetResource))
-
-	return ctrl.Result{}, nil
-}
-
-// ========================================
-// CheckCooldownActive checks if cooldown is active for a target resource
-// BR-WE-009: Cooldown Period is Configurable
-// Returns (remaining duration, is active)
-// currentWFEName format: "namespace/name" to uniquely identify the current WFE
-// ========================================
-func (r *WorkflowExecutionReconciler) CheckCooldownActive(ctx context.Context, targetResource, currentWFEKey string) (time.Duration, bool) {
-	logger := log.FromContext(ctx)
-
-	// Get cooldown period (use default if not set)
-	cooldown := r.CooldownPeriod
-	if cooldown == 0 {
-		cooldown = DefaultCooldownPeriod
-	}
-
-	// Query all WorkflowExecutions with the same targetResource.
-	// Retry up to 3 times with short backoff to tolerate transient informer/API
-	// pressure in resource-constrained environments (e.g., Kind CI clusters).
-	// Only fail-closed if ALL attempts fail (DD-WE-001 compliance).
-	wfeList := &workflowexecutionv1alpha1.WorkflowExecutionList{}
-	const maxAttempts = 3
-	var listErr error
-	for attempt := 1; attempt <= maxAttempts; attempt++ {
-		listErr = r.List(ctx, wfeList, client.MatchingFields{"spec.targetResource": targetResource})
-		if listErr == nil {
-			break
-		}
-		if attempt < maxAttempts {
-			time.Sleep(50 * time.Millisecond)
-		}
-	}
-	if listErr != nil {
-		logger.Error(listErr, "Failed to list WorkflowExecutions for cooldown check after retries — failing closed",
-			"targetResource", targetResource, "attempts", maxAttempts)
-		return cooldown, true
-	}
-
-	// Find any completed/failed WFE still within cooldown period
-	now := time.Now()
-	for i := range wfeList.Items {
-		otherWFE := &wfeList.Items[i]
-
-		// Skip the current WFE (don't check cooldown against ourselves)
-		otherKey := fmt.Sprintf("%s/%s", otherWFE.Namespace, otherWFE.Name)
-		if otherKey == currentWFEKey {
-			continue
-		}
-
-		// Only check terminal phases (Completed or Failed)
-		if otherWFE.Status.Phase != workflowexecutionv1alpha1.PhaseCompleted &&
-			otherWFE.Status.Phase != workflowexecutionv1alpha1.PhaseFailed {
-			continue
-		}
-
-		// Check if completion time is set
-		if otherWFE.Status.CompletionTime == nil {
-			continue
-		}
-
-		// Calculate elapsed time since completion
-		elapsed := now.Sub(otherWFE.Status.CompletionTime.Time)
-
-		// Check if still within cooldown period
-		if elapsed < cooldown {
-			remaining := cooldown - elapsed
-			logger.V(1).Info("Cooldown active for target resource",
-				"targetResource", targetResource,
-				"blockingWFE", otherKey,
-				"remaining", remaining,
-			)
-			return remaining, true
-		}
-	}
-
-	// No active cooldown found
-	return 0, false
-}
-
-// ========================================
-// ReconcileDelete - Handle deletion with finalizer
-// DD-WE-003: Use deterministic PipelineRun name
-// finalizers-lifecycle.md: Event emission
-// ========================================
-func (r *WorkflowExecutionReconciler) ReconcileDelete(ctx context.Context, wfe *workflowexecutionv1alpha1.WorkflowExecution) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-
-	// Issue #518: Lazy resolution for pre-migration WFEs (non-fatal for delete).
-	if _, engineErr := r.resolveExecutionEngine(ctx, wfe); engineErr != nil {
-		logger.V(1).Info("Could not resolve execution engine during delete (non-fatal, will try fallback)", "error", engineErr)
-	}
-	logger.Info("Reconciling Delete")
-
-	// Check if finalizer is present
-	if !controllerutil.ContainsFinalizer(wfe, FinalizerName) {
-		return ctrl.Result{}, nil
-	}
-
-	// ========================================
-	// Cleanup: Delete execution resource via executor dispatch (BR-WE-014, DD-WE-003)
-	// This ensures cleanup even if ExecutionRef was never set
-	// ========================================
-	if r.ExecutorRegistry != nil {
-		exec, execErr := r.ExecutorRegistry.Get(wfe.Status.ExecutionEngine)
-		if execErr != nil {
-			logger.Error(execErr, "Unknown engine during finalization cleanup, skipping executor cleanup",
-				"engine", wfe.Status.ExecutionEngine)
-		}
-		if exec != nil {
-			logger.Info("Cleaning up execution resource",
-				"engine", exec.Engine(),
-				"namespace", r.ExecutionNamespace,
-			)
-			if err := exec.Cleanup(ctx, wfe, r.ExecutionNamespace); err != nil {
-				logger.Error(err, "Failed to cleanup execution resource during finalization",
-					"engine", exec.Engine())
-				// DD-EVENT-001 v1.1: Emit CleanupFailed event (P4: Error Path)
-				r.Recorder.Event(wfe, corev1.EventTypeWarning, events.EventReasonCleanupFailed,
-					fmt.Sprintf("Cleanup failed during finalization: %v", err))
-				return ctrl.Result{}, err
-			}
-			logger.Info("Finalizer: cleaned up execution resource", "engine", exec.Engine())
-		}
-	} else {
-		// Fallback: inline Tekton cleanup when ExecutorRegistry is not configured
-		prName := weexecutor.ExecutionResourceName(wfe.Spec.TargetResource)
-		pr := &tektonv1.PipelineRun{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      prName,
-				Namespace: r.ExecutionNamespace,
-			},
-		}
-		if err := r.Delete(ctx, pr); err != nil {
-			if !apierrors.IsNotFound(err) {
-				logger.Error(err, "Failed to delete PipelineRun during finalization")
-				return ctrl.Result{}, err
-			}
-		} else {
-			logger.Info("Finalizer: deleted associated PipelineRun", "pipelineRun", prName)
-		}
-	}
-
-	// ========================================
-	// Emit deletion event (finalizers-lifecycle.md)
-	// ========================================
-	r.Recorder.Event(wfe, corev1.EventTypeNormal, events.EventReasonWorkflowExecutionDeleted,
-		fmt.Sprintf("WorkflowExecution cleanup completed (phase: %s)", wfe.Status.Phase))
-
-	// ========================================
-	// Remove Finalizer
-	// ========================================
-	logger.Info("Removing finalizer", "finalizer", FinalizerName)
-	controllerutil.RemoveFinalizer(wfe, FinalizerName)
-	if err := r.Update(ctx, wfe); err != nil {
-		logger.Error(err, "Failed to remove finalizer")
-		return ctrl.Result{}, err
-	}
-
-	logger.Info("WorkflowExecution cleanup complete")
 	return ctrl.Result{}, nil
 }
 
@@ -1117,157 +422,6 @@ func (r *WorkflowExecutionReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrlBuilder.Complete(r)
 }
 
-// handleJobAlreadyExists implements Issue #374 pre-execution cleanup for completed Jobs.
-// When a Job creation fails with AlreadyExists, this method checks if the existing Job
-// is in a terminal state. If so, it cleans up the stale Job and retries creation.
-//
-// Returns:
-//
-//	(createResult, true, false, "") — Job cleaned up and recreation succeeded.
-//	(nil, false, true, "")          — Cleanup accepted but GC pending; caller should requeue.
-//	(nil, false, false, name)       — Lock is valid (Job still running); 4th value is original WFE name from label (Issue #190).
-//	(nil, false, false, "")         — Unrecoverable error or executor type mismatch.
-func (r *WorkflowExecutionReconciler) handleJobAlreadyExists(
-	ctx context.Context,
-	exec weexecutor.Executor,
-	wfe *workflowexecutionv1alpha1.WorkflowExecution,
-	resourceName string,
-	createOpts weexecutor.CreateOptions,
-) (*weexecutor.CreateResult, bool, bool, string) {
-	logger := log.FromContext(ctx)
-
-	jobExec, ok := exec.(*weexecutor.JobExecutor)
-	if !ok {
-		return nil, false, false, ""
-	}
-
-	completed, checkErr := jobExec.IsCompleted(ctx, wfe.Spec.ClusterID, wfe.Spec.TargetResource, r.ExecutionNamespace)
-	if checkErr != nil {
-		logger.V(1).Info("Could not check existing Job state (may have been deleted)", "error", checkErr)
-		return nil, false, false, ""
-	}
-	if !completed {
-		// Issue #190: Lock is valid (Job still running). Try to identify the
-		// original WFE from the Job's label so the caller can classify this
-		// as Deduplicated rather than Unknown.
-		originalWFE := r.getOriginalWFEFromJob(ctx, resourceName)
-		return nil, false, false, originalWFE
-	}
-
-	logger.Info("Stale completed Job detected, cleaning up before retry (Issue #374)",
-		"resource", resourceName)
-
-	cleanupWFE := &workflowexecutionv1alpha1.WorkflowExecution{
-		Spec: workflowexecutionv1alpha1.WorkflowExecutionSpec{
-			TargetResource: wfe.Spec.TargetResource,
-		},
-	}
-	if cleanErr := jobExec.Cleanup(ctx, cleanupWFE, r.ExecutionNamespace); cleanErr != nil {
-		logger.Error(cleanErr, "Failed to clean up completed Job", "resource", resourceName)
-		return nil, false, false, ""
-	}
-
-	retryResult, retryErr := exec.Create(ctx, wfe, r.ExecutionNamespace, createOpts)
-	if retryErr != nil {
-		if apierrors.IsAlreadyExists(retryErr) {
-			logger.Info("Job cleanup accepted but GC pending, will requeue (Issue #383)",
-				"resource", resourceName)
-			return nil, false, true, ""
-		}
-		logger.Error(retryErr, "Failed to create Job after stale cleanup", "resource", resourceName)
-		return nil, false, false, ""
-	}
-
-	logger.Info("Successfully recreated Job after stale cleanup (Issue #374)",
-		"resource", resourceName, "newJob", retryResult.ResourceName)
-	return retryResult, true, false, ""
-}
-
-// getOriginalWFEFromJob fetches the existing Job by resource name and reads the
-// kubernaut.ai/workflow-execution label. Returns empty string if the Job cannot
-// be found or the label is absent (Issue #190).
-func (r *WorkflowExecutionReconciler) getOriginalWFEFromJob(ctx context.Context, resourceName string) string {
-	var job batchv1.Job
-	if err := r.Get(ctx, client.ObjectKey{
-		Name:      resourceName,
-		Namespace: r.ExecutionNamespace,
-	}, &job); err != nil {
-		return ""
-	}
-	if job.Labels == nil {
-		return ""
-	}
-	return job.Labels["kubernaut.ai/workflow-execution"]
-}
-
-// ========================================
-// HandleAlreadyExists handles the race condition where PipelineRun already exists
-// DD-WE-003: Layer 2 - Execution-time collision handling (not routing)
-// V1.0: Fails WFE if race condition detected (RO should have prevented this)
-// ========================================
-func (r *WorkflowExecutionReconciler) HandleAlreadyExists(ctx context.Context, wfe *workflowexecutionv1alpha1.WorkflowExecution, resourceName string, err error) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-
-	existingPR := &tektonv1.PipelineRun{}
-	if getErr := r.Get(ctx, client.ObjectKey{
-		Name:      resourceName,
-		Namespace: r.ExecutionNamespace,
-	}, existingPR); getErr != nil {
-		logger.Error(getErr, "Failed to get existing PipelineRun", "name", resourceName)
-		markErr := r.MarkFailedWithReason(ctx, wfe, "RaceConditionError", fmt.Sprintf("PipelineRun already exists but failed to verify ownership: %v", getErr))
-		return ctrl.Result{}, markErr
-	}
-
-	if existingPR.Labels != nil &&
-		existingPR.Labels["kubernaut.ai/workflow-execution"] == wfe.Name &&
-		existingPR.Labels["kubernaut.ai/source-namespace"] == wfe.Namespace {
-		logger.Info("PipelineRun already exists and is ours, continuing", "name", resourceName)
-
-		now := metav1.Now()
-		if err := r.StatusManager.AtomicStatusUpdate(ctx, wfe, func() error {
-			if err := r.PhaseManager.TransitionTo(wfe, wephase.Running); err != nil {
-				return fmt.Errorf("failed to transition to Running in HandleAlreadyExists: %w", err)
-			}
-
-			wfe.Status.StartTime = &now
-			wfe.Status.ExecutionRef = &corev1.LocalObjectReference{
-				Name: resourceName,
-			}
-
-			weconditions.SetExecutionCreated(wfe, true,
-				weconditions.ReasonExecutionCreated,
-				fmt.Sprintf("PipelineRun %s already exists (race condition)", resourceName))
-
-			return nil
-		}); err != nil {
-			return ctrl.Result{}, fmt.Errorf("failed to update status in HandleAlreadyExists: %w", err)
-		}
-
-		r.Recorder.Event(wfe, corev1.EventTypeNormal, events.EventReasonPipelineRunCreated,
-			fmt.Sprintf("PipelineRun %s/%s (already exists, ours)", r.ExecutionNamespace, resourceName))
-
-		return ctrl.Result{RequeueAfter: 10 * time.Second}, nil
-	}
-
-	// Issue #190: Another WFE created this PipelineRun — classify as Deduplicated
-	// when the original WFE can be identified from labels.
-	originalWFE := existingPR.Labels["kubernaut.ai/workflow-execution"]
-	logger.Info("Execution-time collision: PipelineRun created by another WFE",
-		"prName", resourceName,
-		"existingWFE", originalWFE,
-		"targetResource", wfe.Spec.TargetResource,
-	)
-
-	if originalWFE != "" {
-		markErr := r.MarkFailedAsDeduplicated(ctx, wfe, originalWFE)
-		return ctrl.Result{}, markErr
-	}
-
-	markErr := r.MarkFailedWithReason(ctx, wfe, "Unknown",
-		fmt.Sprintf("PipelineRun '%s' already exists for target resource (owner label missing)", resourceName))
-	return ctrl.Result{}, markErr
-}
-
 // ========================================
 // FindWFEForOwnedResource maps owned resource events (PipelineRun, Job) to WorkflowExecution reconcile requests.
 // Both executors label their resources with kubernaut.ai/workflow-execution and kubernaut.ai/source-namespace.
@@ -1291,531 +445,6 @@ func (r *WorkflowExecutionReconciler) FindWFEForOwnedResource(ctx context.Contex
 			Namespace: sourceNS,
 		},
 	}}
-}
-
-// ========================================
-// Day 5: Status Synchronization Methods
-// ========================================
-
-// BuildPipelineRunStatusSummary creates a lightweight status summary from PipelineRun
-// Provides visibility into task progress during execution (v3.2)
-func (r *WorkflowExecutionReconciler) BuildPipelineRunStatusSummary(ctx context.Context, pr *tektonv1.PipelineRun) *workflowexecutionv1alpha1.ExecutionStatusSummary {
-	summary := &workflowexecutionv1alpha1.ExecutionStatusSummary{
-		Status: corev1.ConditionUnknown,
-	}
-
-	// Extract task counts from ChildReferences
-	summary.TotalTasks = len(pr.Status.ChildReferences)
-
-	// Count TaskRuns with ConditionSucceeded True (completed tasks)
-	for _, ref := range pr.Status.ChildReferences {
-		if ref.Kind != "TaskRun" {
-			continue
-		}
-		var tr tektonv1.TaskRun
-		if err := r.Get(ctx, client.ObjectKey{
-			Name:      ref.Name,
-			Namespace: pr.Namespace,
-		}, &tr); err != nil {
-			if apierrors.IsNotFound(err) {
-				continue
-			}
-			continue
-		}
-		cond := tr.Status.GetCondition(apis.ConditionSucceeded)
-		if cond != nil && cond.IsTrue() {
-			summary.CompletedTasks++
-		}
-	}
-
-	// Get Succeeded condition
-	succeededCond := pr.Status.GetCondition(apis.ConditionSucceeded)
-	if succeededCond != nil {
-		summary.Status = corev1.ConditionStatus(succeededCond.Status)
-		summary.Reason = succeededCond.Reason
-		summary.Message = succeededCond.Message
-	}
-
-	return summary
-}
-
-// MarkCompleted transitions WFE to Completed phase
-// Calculates Duration from StartTime to CompletionTime (v3.2)
-// Day 6 Extension (BR-WE-012): Resets ConsecutiveFailures counter
-// Records metrics per BR-WE-008 (Day 7)
-func (r *WorkflowExecutionReconciler) MarkCompleted(ctx context.Context, wfe *workflowexecutionv1alpha1.WorkflowExecution, pr *tektonv1.PipelineRun, summary ...*workflowexecutionv1alpha1.ExecutionStatusSummary) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-	logger.Info("Marking WorkflowExecution as Completed")
-
-	// ========================================
-	// DD-RO-002 Phase 3: Counter Reset Removed (Dec 19, 2025)
-	// RO resets RR.Status.ConsecutiveFailureCount on successful remediation
-	// WE no longer tracks routing state
-	// ========================================
-
-	// Calculate duration for use in atomic update
-	now := metav1.Now()
-	var completionTime *metav1.Time
-	if pr != nil && pr.Status.CompletionTime != nil {
-		completionTime = pr.Status.CompletionTime
-	} else {
-		completionTime = &now
-	}
-
-	var durationVal *metav1.Duration
-	var durationSeconds float64
-	if wfe.Status.StartTime != nil && completionTime != nil {
-		d := completionTime.Sub(wfe.Status.StartTime.Time).Round(time.Second)
-		durationVal = &metav1.Duration{Duration: d}
-		durationSeconds = d.Seconds()
-	}
-
-	// ========================================
-	// DD-PERF-001: ATOMIC STATUS UPDATE
-	// Consolidates phase transition + conditions into single API call
-	// BEFORE: 2 API calls (phase update + conditions update)
-	// AFTER: 1 atomic API call (50% reduction)
-	// ========================================
-	if err := r.StatusManager.AtomicStatusUpdate(ctx, wfe, func() error {
-		// Set phase (P0: Phase State Machine)
-		if err := r.PhaseManager.TransitionTo(wfe, wephase.Completed); err != nil {
-			return fmt.Errorf("failed to transition to Completed: %w", err)
-		}
-
-		// Set completion time
-		wfe.Status.CompletionTime = completionTime
-
-		// Set duration
-		wfe.Status.Duration = durationVal
-
-		// Issue #118 Gap 4: persist ExecutionStatus inside callback (survives refetch)
-		if len(summary) > 0 && summary[0] != nil {
-			wfe.Status.ExecutionStatus = summary[0]
-		}
-
-		// BR-WE-006: Set TektonPipelineComplete condition
-		weconditions.SetExecutionComplete(wfe, true,
-			weconditions.ReasonExecutionSucceeded,
-			fmt.Sprintf("All tasks completed successfully in %s", wfe.Status.Duration))
-
-		// Issue #79 Phase 7b: Set Ready condition on terminal transitions
-		weconditions.SetReady(wfe, true, weconditions.ReasonReady, "Workflow execution completed")
-
-		// Day 8: Record audit event for workflow completion (BR-WE-005, ADR-032)
-		if err := r.AuditManager.RecordWorkflowCompleted(ctx, wfe); err != nil {
-			logger.V(1).Info("Failed to record workflow.completed audit event", "error", err)
-			weconditions.SetAuditRecorded(wfe, false,
-				weconditions.ReasonAuditFailed,
-				fmt.Sprintf("Failed to record audit event: %v", err))
-		} else {
-			weconditions.SetAuditRecorded(wfe, true,
-				weconditions.ReasonAuditSucceeded,
-				"Audit event workflow.completed recorded to DataStorage")
-		}
-
-		return nil
-	}); err != nil {
-		logger.Error(err, "Failed to atomically update status to Completed")
-		return ctrl.Result{}, err
-	}
-
-	// Day 7: Record metrics (BR-WE-008)
-	// DD-METRICS-001: Use injected metrics instead of global function
-	if r.Metrics != nil {
-		r.Metrics.RecordWorkflowCompletion(durationSeconds)
-	}
-
-	// V1.0: Consecutive failures gauge removed - RO handles routing (DD-RO-002)
-
-	// Emit event
-	r.Recorder.Event(wfe, corev1.EventTypeNormal, events.EventReasonWorkflowCompleted,
-		fmt.Sprintf("Workflow %s completed successfully in %s", wfe.Spec.WorkflowRef.WorkflowID, wfe.Status.Duration))
-
-	// DD-EVENT-001 v1.1: PhaseTransition breadcrumb for Running → Completed
-	r.emitPhaseTransition(wfe, "Running", "Completed")
-
-	logger.Info("WorkflowExecution completed (atomic status update)")
-
-	// Issue #375: Schedule requeue so ReconcileTerminal runs after cooldown to
-	// delete the execution resource (Job/PipelineRun) and release the target lock.
-	// Without this, GenerationChangedPredicate blocks the status-only update event,
-	// and ReconcileTerminal is never called.
-	cooldown := r.CooldownPeriod
-	if cooldown == 0 {
-		cooldown = DefaultCooldownPeriod
-	}
-	return ctrl.Result{RequeueAfter: cooldown}, nil
-}
-
-// MarkFailed transitions WFE to Failed phase with FailureDetails
-// Extracts failure information from PipelineRun (v3.2)
-// Day 6 Extension (BR-WE-012): Handles exponential backoff for pre-execution failures
-// Records metrics per BR-WE-008 (Day 7)
-func (r *WorkflowExecutionReconciler) MarkFailed(ctx context.Context, wfe *workflowexecutionv1alpha1.WorkflowExecution, pr *tektonv1.PipelineRun, summary ...*workflowexecutionv1alpha1.ExecutionStatusSummary) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
-	logger.Info("Marking WorkflowExecution as Failed")
-
-	now := metav1.Now()
-	var durationVal *metav1.Duration
-	var durationSeconds float64
-	if wfe.Status.StartTime != nil {
-		d := now.Sub(wfe.Status.StartTime.Time).Round(time.Second)
-		durationVal = &metav1.Duration{Duration: d}
-		durationSeconds = d.Seconds()
-	}
-
-	// Extract failure details (Day 7: includes TaskRun-specific fields, Day 6 Extension: WasExecutionFailure)
-	failureDetails := r.ExtractFailureDetails(ctx, pr, wfe.Status.StartTime)
-
-	// BR-WE-015: For non-Tekton executors (ansible, job), the PipelineRun is nil and
-	// ExtractFailureDetails defaults to "Unknown". Override with the executor's summary
-	// which carries the correct engine-specific reason and message.
-	if pr == nil && len(summary) > 0 && summary[0] != nil && failureDetails != nil {
-		if summary[0].Reason != "" {
-			failureDetails.Reason = mapExecutorReasonToCRDEnum(summary[0].Reason)
-		}
-		if summary[0].Message != "" {
-			failureDetails.Message = summary[0].Message
-		}
-		failureDetails.WasExecutionFailure = true
-	}
-
-	// Generate natural language summary
-	if failureDetails != nil {
-		failureDetails.NaturalLanguageSummary = r.GenerateNaturalLanguageSummary(wfe, failureDetails)
-	}
-
-	// Determine condition values
-	failureReason := weconditions.ReasonExecutionFailed
-	failureMessage := "Pipeline execution failed"
-	if failureDetails != nil {
-		// Map WE failure reasons to condition reasons
-		switch failureDetails.Reason {
-		case "TaskFailed":
-			failureReason = weconditions.ReasonTaskFailed
-			failureMessage = failureDetails.Message
-		case "DeadlineExceeded":
-			failureReason = weconditions.ReasonDeadlineExceeded
-			failureMessage = "Pipeline exceeded timeout deadline"
-		case "OOMKilled":
-			failureReason = weconditions.ReasonOOMKilled
-			failureMessage = "Pipeline task killed due to out of memory"
-		default:
-			failureMessage = failureDetails.Message
-		}
-	}
-
-	// ========================================
-	// DD-RO-002 Phase 3: Routing Logic Removed (Dec 19, 2025)
-	// WE is now a pure executor - no routing decisions
-	// RO tracks ConsecutiveFailureCount and NextAllowedExecution in RR.Status
-	// RO makes ALL routing decisions BEFORE creating WFE
-	// ========================================
-	logger.V(1).Info("Workflow execution failed - routing handled by RO",
-		"wasExecutionFailure", failureDetails != nil && failureDetails.WasExecutionFailure)
-
-	// ========================================
-	// DD-PERF-001: ATOMIC STATUS UPDATE
-	// Consolidates phase transition + conditions into single API call
-	// ========================================
-	if err := r.StatusManager.AtomicStatusUpdate(ctx, wfe, func() error {
-		// Set phase (P0: Phase State Machine)
-		if err := r.PhaseManager.TransitionTo(wfe, wephase.Failed); err != nil {
-			return fmt.Errorf("failed to transition to Failed: %w", err)
-		}
-
-		// Set completion time
-		wfe.Status.CompletionTime = &now
-
-		// Set duration
-		wfe.Status.Duration = durationVal
-
-		// Set failure details
-		wfe.Status.FailureDetails = failureDetails
-
-		// Issue #118 Gap 4: persist ExecutionStatus inside callback (survives refetch)
-		if len(summary) > 0 && summary[0] != nil {
-			wfe.Status.ExecutionStatus = summary[0]
-		}
-
-		// BR-WE-006: Set TektonPipelineComplete condition to False
-		weconditions.SetExecutionComplete(wfe, false,
-			failureReason,
-			failureMessage)
-
-		// Issue #79 Phase 7b: Set Ready condition on terminal transitions
-		weconditions.SetReady(wfe, false, weconditions.ReasonNotReady, "Workflow execution failed")
-
-		// Day 8: Record audit event for workflow failure (BR-WE-005, ADR-032)
-		// Uses Audit Manager (P3: Audit Manager pattern)
-		if err := r.AuditManager.RecordWorkflowFailed(ctx, wfe); err != nil {
-			logger.V(1).Info("Failed to record workflow.failed audit event", "error", err)
-			weconditions.SetAuditRecorded(wfe, false,
-				weconditions.ReasonAuditFailed,
-				fmt.Sprintf("Failed to record audit event: %v", err))
-		} else {
-			weconditions.SetAuditRecorded(wfe, true,
-				weconditions.ReasonAuditSucceeded,
-				"Audit event workflow.failed recorded to DataStorage")
-		}
-
-		return nil
-	}); err != nil {
-		logger.Error(err, "Failed to atomically update status to Failed")
-		return ctrl.Result{}, err
-	}
-
-	// Day 7: Record metrics (BR-WE-008)
-	// DD-METRICS-001: Use injected metrics instead of global function
-	if r.Metrics != nil {
-		r.Metrics.RecordWorkflowFailure(durationSeconds)
-	}
-
-	// V1.0: Consecutive failures gauge removed - RO handles routing (DD-RO-002)
-
-	// Emit event
-	reason := "Unknown"
-	if wfe.Status.FailureDetails != nil {
-		reason = wfe.Status.FailureDetails.Reason
-	}
-	r.Recorder.Event(wfe, corev1.EventTypeWarning, events.EventReasonWorkflowFailed,
-		fmt.Sprintf("Workflow %s failed: %s", wfe.Spec.WorkflowRef.WorkflowID, reason))
-
-	// DD-EVENT-001 v1.1: PhaseTransition breadcrumb for Running → Failed
-	r.emitPhaseTransition(wfe, "Running", "Failed")
-
-	// Issue #375: Schedule requeue so ReconcileTerminal runs after cooldown to
-	// delete the execution resource (Job/PipelineRun) and release the target lock.
-	cooldown := r.CooldownPeriod
-	if cooldown == 0 {
-		cooldown = DefaultCooldownPeriod
-	}
-	return ctrl.Result{RequeueAfter: cooldown}, nil
-}
-
-// ========================================
-// MarkFailedWithReason - Handle pre-execution failures
-// Used for validation errors, configuration errors before PipelineRun creation
-// ========================================
-func (r *WorkflowExecutionReconciler) MarkFailedWithReason(ctx context.Context, wfe *workflowexecutionv1alpha1.WorkflowExecution, reason, message string) error {
-	logger := log.FromContext(ctx)
-	logger.Info("Marking WorkflowExecution as Failed (pre-execution)",
-		"reason", reason,
-		"message", message,
-	)
-
-	now := metav1.Now()
-	failureDetails := &workflowexecutionv1alpha1.FailureDetails{
-		Reason:              reason,
-		Message:             message,
-		FailedAt:            now,
-		WasExecutionFailure: false,
-	}
-	failureDetails.NaturalLanguageSummary = r.GenerateNaturalLanguageSummary(wfe, failureDetails)
-
-	conditionReason := weconditions.ReasonExecutionCreationFailed
-	switch reason {
-	case "QuotaExceeded":
-		conditionReason = weconditions.ReasonQuotaExceeded
-	case "PermissionDenied", "RBACDenied":
-		conditionReason = weconditions.ReasonRBACDenied
-	case "ImagePullFailed":
-		conditionReason = weconditions.ReasonImagePullFailed
-	}
-
-	condMsg := fmt.Sprintf("Failed to create PipelineRun: %s", message)
-	if err := r.markFailedInternal(ctx, wfe, failureDetails, conditionReason, condMsg, &now, nil); err != nil {
-		return err
-	}
-
-	r.Recorder.Event(wfe, corev1.EventTypeWarning, events.EventReasonWorkflowFailed,
-		fmt.Sprintf("Pre-execution failure: %s - %s", reason, message))
-	r.emitPhaseTransition(wfe, "Pending", "Failed")
-
-	logger.Info("WorkflowExecution failed with reason (atomic status update)", "reason", reason)
-	return nil
-}
-
-// MarkFailedAsDeduplicated marks a WFE as failed due to an execution-time resource
-// collision with another WFE. Sets FailureDetails.Reason = Deduplicated and
-// DeduplicatedBy = originalWFE inside the AtomicStatusUpdate closure to satisfy
-// the M5 constraint (refetch-safe atomic writes). Issue #190.
-func (r *WorkflowExecutionReconciler) MarkFailedAsDeduplicated(ctx context.Context, wfe *workflowexecutionv1alpha1.WorkflowExecution, originalWFE string) error {
-	logger := log.FromContext(ctx)
-	logger.Info("Marking WorkflowExecution as Failed (deduplicated)", "originalWFE", originalWFE)
-
-	now := metav1.Now()
-	failureDetails := &workflowexecutionv1alpha1.FailureDetails{
-		Reason:              workflowexecutionv1alpha1.FailureReasonDeduplicated,
-		Message:             fmt.Sprintf("Execution resource already exists, owned by WorkflowExecution %s", originalWFE),
-		FailedAt:            now,
-		WasExecutionFailure: false,
-	}
-	failureDetails.NaturalLanguageSummary = r.GenerateNaturalLanguageSummary(wfe, failureDetails)
-
-	condMsg := fmt.Sprintf("Execution resource collision (deduplicated by %s)", originalWFE)
-	extraUpdates := func() {
-		wfe.Status.DeduplicatedBy = originalWFE
-	}
-	if err := r.markFailedInternal(ctx, wfe, failureDetails, weconditions.ReasonExecutionCreationFailed, condMsg, &now, extraUpdates); err != nil {
-		return err
-	}
-
-	r.Recorder.Event(wfe, corev1.EventTypeWarning, events.EventReasonWorkflowFailed,
-		fmt.Sprintf("Deduplicated: execution resource collision with %s", originalWFE))
-	r.emitPhaseTransition(wfe, "Pending", "Failed")
-	return nil
-}
-
-// markFailedInternal is the shared core for MarkFailedWithReason and
-// MarkFailedAsDeduplicated. It performs the AtomicStatusUpdate with phase
-// transition, completion time, failure details, conditions, audit, and metrics.
-// The optional extraUpdates callback runs inside the atomic closure to set
-// method-specific fields (e.g., DeduplicatedBy for the M5 constraint).
-func (r *WorkflowExecutionReconciler) markFailedInternal(
-	ctx context.Context,
-	wfe *workflowexecutionv1alpha1.WorkflowExecution,
-	failureDetails *workflowexecutionv1alpha1.FailureDetails,
-	conditionReason, conditionMessage string,
-	now *metav1.Time,
-	extraUpdates func(),
-) error {
-	logger := log.FromContext(ctx)
-
-	if err := r.StatusManager.AtomicStatusUpdate(ctx, wfe, func() error {
-		if err := r.PhaseManager.TransitionTo(wfe, wephase.Failed); err != nil {
-			return fmt.Errorf("failed to transition to Failed: %w", err)
-		}
-
-		wfe.Status.CompletionTime = now
-		wfe.Status.FailureDetails = failureDetails
-
-		if extraUpdates != nil {
-			extraUpdates()
-		}
-
-		weconditions.SetExecutionCreated(wfe, false, conditionReason, conditionMessage)
-		weconditions.SetReady(wfe, false, weconditions.ReasonNotReady, "Workflow execution failed")
-
-		if err := r.AuditManager.RecordWorkflowFailed(ctx, wfe); err != nil {
-			logger.V(1).Info("Failed to record workflow.failed audit event", "error", err)
-			weconditions.SetAuditRecorded(wfe, false,
-				weconditions.ReasonAuditFailed,
-				fmt.Sprintf("Failed to record audit event: %v", err))
-		} else {
-			weconditions.SetAuditRecorded(wfe, true,
-				weconditions.ReasonAuditSucceeded,
-				"Audit event workflow.failed recorded to DataStorage")
-		}
-
-		return nil
-	}); err != nil {
-		logger.Error(err, "Failed to atomically update status to Failed")
-		return err
-	}
-
-	if r.Metrics != nil {
-		r.Metrics.RecordWorkflowFailure(0)
-	}
-	return nil
-}
-
-// ========================================
-// Helper Functions for Status Updates
-// ========================================
-
-// updateStatus is a helper that updates the WFE status with consistent error handling
-func (r *WorkflowExecutionReconciler) updateStatus(
-	ctx context.Context,
-	wfe *workflowexecutionv1alpha1.WorkflowExecution,
-	operation string,
-) error {
-	logger := log.FromContext(ctx)
-
-	if err := r.Status().Update(ctx, wfe); err != nil {
-		logger.Error(err, "Failed to update status", "operation", operation)
-		return err
-	}
-	return nil
-}
-
-// ========================================
-// Issue #518: Runtime Execution Engine Resolution
-// ========================================
-
-// resolveWorkflowCatalog fetches all workflow metadata from the DS catalog in a
-// single GetWorkflowByID call (Issue #650). Consolidates resolveExecutionEngine,
-// resolveExecutionBundle, resolveDependencies, and GetWorkflowEngineConfig.
-// Idempotent: returns nil immediately if the engine is already resolved.
-func (r *WorkflowExecutionReconciler) resolveWorkflowCatalog(ctx context.Context, wfe *workflowexecutionv1alpha1.WorkflowExecution) (*weclient.WorkflowCatalogMetadata, error) {
-	if wfe.Status.ExecutionEngine != "" {
-		return nil, nil
-	}
-
-	if r.WorkflowQuerier == nil {
-		return nil, fmt.Errorf("DataStorage workflow querier not available — cannot resolve workflow catalog for %s", wfe.Spec.WorkflowRef.WorkflowID)
-	}
-
-	workflowID := wfe.Spec.WorkflowRef.WorkflowID
-	if workflowID == "" {
-		return nil, fmt.Errorf("workflowRef.workflowId is empty — cannot resolve workflow catalog")
-	}
-
-	logger := log.FromContext(ctx)
-
-	meta, err := r.WorkflowQuerier.ResolveWorkflowCatalogMetadata(ctx, workflowID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to resolve workflow catalog from DS for workflow %s: %w", workflowID, err)
-	}
-
-	if meta.ExecutionEngine == "" {
-		label := workflowID
-		if meta.WorkflowName != "" {
-			label = fmt.Sprintf("%s - %s", meta.WorkflowName, workflowID)
-		}
-		return nil, fmt.Errorf("no engine defined in remediation workflow %s", label)
-	}
-
-	wfe.Status.ExecutionEngine = meta.ExecutionEngine
-	wfe.Status.ServiceAccountName = meta.ServiceAccountName
-
-	// WE-H1: In-memory spec mutation is acceptable here because ResolveWorkflowCatalogMetadata
-	// is called on every Pending-phase reconcile. If the controller requeues in Pending,
-	// the catalog lookup re-applies the override. Once the WFE leaves Pending, the bundle
-	// value is already consumed (execution resource created).
-	if meta.ExecutionBundle != "" {
-		if wfe.Spec.WorkflowRef.ExecutionBundle != meta.ExecutionBundle {
-			logger.Info("Overriding execution bundle from DS catalog",
-				"specBundle", wfe.Spec.WorkflowRef.ExecutionBundle,
-				"catalogBundle", meta.ExecutionBundle,
-				"workflowID", workflowID,
-			)
-		}
-		wfe.Spec.WorkflowRef.ExecutionBundle = meta.ExecutionBundle
-		if meta.ExecutionBundleDigest != "" {
-			wfe.Spec.WorkflowRef.ExecutionBundleDigest = meta.ExecutionBundleDigest
-		}
-	}
-
-	if wfe.Spec.WorkflowRef.EngineConfig == nil && meta.EngineConfig != nil {
-		logger.Info("Resolved engineConfig from DS catalog",
-			"workflowID", workflowID,
-			"engine", meta.ExecutionEngine)
-		wfe.Spec.WorkflowRef.EngineConfig = &apiextensionsv1.JSON{Raw: meta.EngineConfig}
-	}
-
-	return meta, nil
-}
-
-// resolveExecutionEngine returns the cached execution engine from the WFE status.
-// In non-Pending phases the engine was already resolved during Pending and persisted
-// to wfe.Status.ExecutionEngine. Returns an error only if the engine is missing,
-// which indicates a programming error (Pending handler should have set it).
-func (r *WorkflowExecutionReconciler) resolveExecutionEngine(_ context.Context, wfe *workflowexecutionv1alpha1.WorkflowExecution) (string, error) {
-	if wfe.Status.ExecutionEngine != "" {
-		return wfe.Status.ExecutionEngine, nil
-	}
-	return "", fmt.Errorf("execution engine not resolved for WFE %s/%s — expected to be set during Pending phase", wfe.Namespace, wfe.Name)
 }
 
 // ========================================
